+++ src/main/java/com/vivern/arpg/blocks/AshBlock.java
@@ -52,7 +52,7 @@ public class AshBlock extends BlockFalling {
    @Override
    @Nullable
    public AxisAlignedBB getCollisionBoundingBox(IBlockState blockState, IBlockAccess worldIn, BlockPos pos) {
-      int i = (Integer)blockState.getValue(LAYERS);
+      int i = blockState.getValue(LAYERS);
       AxisAlignedBB axisalignedbb = blockState.getBoundingBox(worldIn, pos);
       return i == 8
          ? SNOW_AABB[8]
@@ -73,7 +73,7 @@ public class AshBlock extends BlockFalling {
 
    @Override
    public boolean isPassable(IBlockAccess worldIn, BlockPos pos) {
-      return (Integer)worldIn.getBlockState(pos).getValue(LAYERS) < 5;
+      return worldIn.getBlockState(pos).getValue(LAYERS) < 5;
    }
 
    @Override
@@ -83,13 +83,13 @@ public class AshBlock extends BlockFalling {
 
    @Override
    public int quantityDropped(IBlockState state, int fortune, Random rand) {
-      int layersAmount = (Integer)state.getValue(LAYERS);
+      int layersAmount = state.getValue(LAYERS);
       return MathHelper.ceil(rand.nextFloat() * layersAmount / 2.0F);
    }
 
    @Override
    public boolean isFullCube(IBlockState state) {
-      return (Integer)state.getValue(LAYERS) == 8;
+      return state.getValue(LAYERS) == 8;
    }
 
    @Override
@@ -109,7 +109,7 @@ public class AshBlock extends BlockFalling {
 
    @Override
    public int getMetaFromState(IBlockState state) {
-      return (Integer)state.getValue(LAYERS);
+      return state.getValue(LAYERS);
    }
 
    @Override
@@ -122,7 +122,7 @@ public class AshBlock extends BlockFalling {
       if (replaced.getBlock() == this) {
          worldIn.setBlockState(
             pos,
-            placed.withProperty(LAYERS, Math.min(8, (Integer)replaced.getValue(LAYERS) + (Integer)placed.getValue(LAYERS)))
+            placed.withProperty(LAYERS, Math.min(8, replaced.getValue(LAYERS) + placed.getValue(LAYERS)))
                .withProperty(ISFALLING, false)
          );
       } else {
@@ -138,7 +138,7 @@ public class AshBlock extends BlockFalling {
    }
 
    private void checkFallable(World worldIn, BlockPos pos) {
-      if (worldIn.getBlockState(pos.down()).getBlock() == this && (Integer)worldIn.getBlockState(pos.down()).getValue(LAYERS) < 8
+      if (worldIn.getBlockState(pos.down()).getBlock() == this && worldIn.getBlockState(pos.down()).getValue(LAYERS) < 8
          || (worldIn.isAirBlock(pos.down()) || canFallThrough(worldIn.getBlockState(pos.down()))) && pos.getY() >= 0) {
          int i = 32;
          if (fallInstantly || !worldIn.isAreaLoaded(pos.add(-32, -32, -32), pos.add(32, 32, 32))) {
@@ -175,7 +175,7 @@ public class AshBlock extends BlockFalling {
 
    @Override
    public boolean isSideSolid(IBlockState state, IBlockAccess world, BlockPos pos, EnumFacing side) {
-      return side == EnumFacing.DOWN || (Integer)state.getValue(LAYERS) == 8;
+      return side == EnumFacing.DOWN || state.getValue(LAYERS) == 8;
    }
 
    @Override
