+++ src/main/java/com/vivern/arpg/blocks/CaveCrystal.java
@@ -110,7 +110,7 @@ public class CaveCrystal extends Block {
 
    @Override
    public int getMetaFromState(IBlockState state) {
-      return (Integer)state.getValue(TYPE);
+      return state.getValue(TYPE);
    }
 
    @Override
