+++ src/main/java/com/vivern/arpg/blocks/GlowingCaveCrystal.java
@@ -116,7 +116,7 @@ public class GlowingCaveCrystal extends Block {
 
    @Override
    public int getMetaFromState(IBlockState state) {
-      return (Integer)state.getValue(TYPE);
+      return state.getValue(TYPE);
    }
 
    @Override
