+++ src/main/java/com/vivern/arpg/blocks/MagicOrnament.java
@@ -24,7 +24,7 @@ public class MagicOrnament extends BlockBlock {
 
    @Override
    public int getMetaFromState(IBlockState state) {
-      return (Integer)state.getValue(TYPE);
+      return state.getValue(TYPE);
    }
 
    @Override
