+++ src/main/java/com/vivern/arpg/blocks/Seagrass.java
@@ -133,7 +133,7 @@ public class Seagrass extends Block {
       for (EnumFacing facing : EnumFacing.VALUES) {
          BlockPos poss = pos.offset(facing);
          IBlockState state2 = world.getBlockState(poss);
-         if (state2.getBlock() == Blocks.WATER && (Integer)state2.getValue(BlockStaticLiquid.LEVEL) == 0
+         if (state2.getBlock() == Blocks.WATER && state2.getValue(BlockStaticLiquid.LEVEL) == 0
             || state2.getBlock() == this) {
             if (++count >= 2) {
                return true;
