+++ src/main/java/com/vivern/arpg/blocks/SeleniteCrystal.java
@@ -108,7 +108,7 @@ public class SeleniteCrystal extends Block {
 
    @Override
    public int getMetaFromState(IBlockState state) {
-      return (Integer)state.getValue(TYPE);
+      return state.getValue(TYPE);
    }
 
    @Override
