+++ src/main/java/com/vivern/arpg/blocks/TideBeacon.java
@@ -113,7 +113,7 @@ public class TideBeacon extends Block implements IBlockHardBreak {
       for (EnumFacing facing : EnumFacing.VALUES) {
          BlockPos poss = pos.offset(facing);
          IBlockState state2 = world.getBlockState(poss);
-         if (state2.getBlock() == Blocks.WATER && (Integer)state2.getValue(BlockStaticLiquid.LEVEL) == 0) {
+         if (state2.getBlock() == Blocks.WATER && state2.getValue(BlockStaticLiquid.LEVEL) == 0) {
             if (++count >= 2) {
                return true;
             }
